% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{availableDatasets}
\alias{availableDatasets}
\title{List the available datasets}
\usage{
availableDatasets()
}
\value{
A character vector with the available datasets.
}
\description{
List the available datasets
}
\examples{
library(OmopDatasets)

availableDatasets()

}
