% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{datasetNameDoc}
\alias{datasetNameDoc}
\title{Helper for consistent documentation of \code{datasetName}.}
\arguments{
\item{datasetName}{Name of the dataset. See \code{availableDatasets()} for
possibilities.}
}
\description{
Helper for consistent documentation of \code{datasetName}.
}
\keyword{internal}
