% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datasetsFolder}
\alias{datasetsFolder}
\title{Check or set the datasets Folder}
\usage{
datasetsFolder(path = NULL)
}
\arguments{
\item{path}{Path to a folder to store the synthetic datasets. If NULL the
current OMOP_DATASETS_FOLDER is returned.}
}
\value{
The dataset folder.
}
\description{
Check or set the datasets Folder
}
\examples{
\donttest{
datasetsFolder()
datasetsFolder(file.path(tempdir(), "OMOP_DATASETS"))
datasetsFolder()
}

}
