% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{datasetStatus}
\alias{datasetStatus}
\title{Check the availability of the OMOP CDM datasets.}
\usage{
datasetStatus()
}
\value{
A message with the availability of the datasets.
}
\description{
Check the availability of the OMOP CDM datasets.
}
\examples{
datasetStatus()
}
