% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download an OMOP Synthetic dataset.}
\usage{
downloadDataset(
  datasetName = "GiBleed",
  path = datasetsFolder(),
  overwrite = FALSE
)
}
\arguments{
\item{datasetName}{Name of the dataset. See \code{availableDatasets()} for
possibilities.}

\item{path}{Path where to download the dataset.}

\item{overwrite}{Whether to overwrite the dataset if it is already
downloaded.}
}
\value{
The path to the downloaded dataset.
}
\description{
Download an OMOP Synthetic dataset.
}
\examples{
\donttest{
library(OmopDatasets)

isDatasetDownloaded("GiBleed")
downloadDataset("GiBleed")
isDatasetDownloaded("GiBleed")
}

}
