% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportDatasetToDuckdb}
\alias{exportDatasetToDuckdb}
\title{Title}
\usage{
exportDatasetToDuckdb(
  path,
  datasetName = "GiBleed",
  cdmSchema = "main",
  cdmPrefix = "",
  writeSchema = "main",
  writePrefix = ""
)
}
\arguments{
\item{writePrefix}{}
}
\description{
Title
}
