% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportDatasetToFiles}
\alias{exportDatasetToFiles}
\title{Export a dataset}
\usage{
exportDatasetToFiles(path, datasetName = "GiBleed", format = "csv")
}
\arguments{
\item{format}{}
}
\description{
Export a dataset
}
