% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isDatasetDownloaded}
\alias{isDatasetDownloaded}
\title{Check if a certain dataset is downloaded.}
\usage{
isDatasetDownloaded(datasetName = "GiBleed", path = datasetsFolder())
}
\arguments{
\item{datasetName}{Name of the dataset. See \code{availableDatasets()} for
possibilities.}

\item{path}{Path where to search for the dataset.}
}
\value{
Whether the dataset is available or not.
}
\description{
Check if a certain dataset is downloaded.
}
