% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopDatasets-package.R
\docType{package}
\name{OmopDatasets-package}
\alias{OmopDatasets}
\alias{OmopDatasets-package}
\title{OmopDatasets: Dataset Manager for OMOP Common Data Model Sample Datasets}
\description{
Download, export and edit synthetic sample datasets mapped to the Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopDatasets/}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
