% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{omopDatasets}
\alias{omopDatasets}
\title{Available OMOP CDM Synthetic Datasets}
\format{
A data frame with 4 variables:
\describe{
\item{dataset_name}{Name of the dataset.}
\item{url}{url to download the dataset.}
\item{cdm_name}{Name of the cdm reference created.}
\item{cdm_version}{OMOP CDM version of the dataset.}
\item{size}{Size in bytes of the dataset.}
\item{size_mb}{Size in Mega bytes of the dataset.}
}
}
\usage{
omopDatasets
}
\description{
There are the OMOP CDM Synthetic datasets that are available to download
using the \code{OmopDatasets} package.
}
\examples{
omopDatasets

}
\keyword{datasets}
