% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cdmFromDataset}
\alias{cdmFromDataset}
\title{Create a \code{duckdb} cdm_reference from a dataset.}
\usage{
cdmFromDataset(
  datasetName = "GiBleed",
  cdmSchema = "main",
  cdmPrefix = "",
  writeSchema = "main",
  writePrefix = ""
)
}
\arguments{
\item{datasetName}{Name of the dataset. See \code{availableDatasets()} for
possibilities.}

\item{cdmSchema}{A string that indicates the cdm schema.}

\item{cdmPrefix}{A string that prefixes the tables in the cdm schema.}

\item{writeSchema}{A string that indicates the write schema.}

\item{writePrefix}{A string that prefixes the tables in the write schema.}
}
\value{
A cdm_reference object.
}
\description{
Note these are temporal cdm objects that won't be accessible after
disconnecting, if you want to create a duckdb database use
\code{exportDatasetToDuckdb()}.
}
\examples{
library(OmopDatasets)

datasetsFolder(tempdir())
downloadDataset(datasetName = "GiBleed")
cdm <- cdmFromDataset(datasetName = "GiBleed")
cdm

}
