% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cdmFromDataset}
\alias{cdmFromDataset}
\title{Title}
\usage{
cdmFromDataset(
  datasetName = "GiBleed",
  cdmSchema = "main",
  cdmPrefix = "",
  writeSchema = "main",
  writePrefix = ""
)
}
\arguments{
\item{writePrefix}{}
}
\description{
Title
}
