% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{schemasDoc}
\alias{schemasDoc}
\title{Helper for consistent documentation of \code{schemas}.}
\arguments{
\item{cdmSchema}{A string that indicates the cdm schema.}

\item{cdmPrefix}{A string that prefixes the tables in the cdm schema.}

\item{writeSchema}{A string that indicates the write schema.}

\item{writePrefix}{A string that prefixes the tables in the write schema.}
}
\description{
Helper for consistent documentation of \code{schemas}.
}
\keyword{internal}
